120V HPWH (#134)
Add a 120V HPWH into OCHRE, based on the model in OS-HPMXL. This model
is based on the Rheem, shared circuit product. I can potentially also
add the dedicated circuit version (higher power, but requires an open
available 15A circuit). We can also allow backup elements like AO Smith
includes.

This also includes adding the ability to model a tempering valve. These
products have an integrated tempering valve, so you can overheat the
tank to increase the stored energy to compensate for the lower heating
rate.

Not all of this is supported in HPXML yet, but that's coming soon. I
want to wait to merge this in until those changes happen so that we can
properly support this going forward, rather than just having something
triggered by temporary flags.

Fixes #133

- [X] Reference the issue your PR is fixing
- [X] Assign at least 1 reviewer for your PR
- [X] Test with run_dwelling.py or other script
- [x] Update documentation as appropriate
- [x] Update changelog as appropriate